+++ Store.Storage/src/BaseClient.cs
@@ -47,7 +47,7 @@ namespace Store {
       /// <returns></returns>
       public abstract List<Record<T>> search(string version, string field, string search);
 
-      public abstract int count(string version, string field = null, string search = null);
+      public abstract long count(string version, string field = null, string search = null);
       
       public U one<U>(string version, string field, string value, Type typeOfParty = null) where U : class {
         var rec = getOneRecord<U>(version, field, value, typeOfParty);
